requirements.txt
fastapi
uvicorn
python-multipart
backcall==0.2.0
branca==0.5.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
decorator==5.1.1
earthengine-api==0.1.316
folium==0.12.1.post1
future==0.18.2
geojson==2.5.0
google-api-core==2.8.2
google-api-python-client==1.12.11
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.1
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.3
httplib2==0.20.4
httplib2shim==0.0.3
idna==3.3
ipython==7.34.0
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
numpy==1.21.6
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
protobuf==4.21.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
pyparsing==3.0.9
requests==2.28.1
rsa==4.8
six==1.16.0
traitlets==5.3.0
uritemplate==3.0.1
urllib3==1.26.10
wcwidth==0.2.5
